requirements.txt
amqp==5.1.1
async-timeout==4.0.3
bidict==0.22.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fsspec==2023.9.2
h11==0.14.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
limits==3.6.0
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
ordered-set==4.1.0
packaging==23.2
prompt-toolkit==3.0.39
psycopg2==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-engineio==4.7.1
python-socketio==5.9.0
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.21
sympy==1.12
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.15.0
wsproto==1.2.0
